Take advantage of remote.Reuse
Each of the remote.Write[Index] calls has to do a Ping and a Token
request, which increases the number of requests we send by quite a lot.

This lets us reuse ping results and authenticated context.

In a follow-up, I will bump apko and pass these options through to that.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>