Make failure to unify a specified package an error.
Currently when we are performing package builds, we will be unable to lock a package version across architectures
and we surface a mediocre warning, and the not-fully-locked configuration creates errors downstream when we try to
create things like SLSA attestations or extract version tags because things are not fully resolved.

By promoting this to an error, we make things fail earlier and with a (now much) clearer error message.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>