Bump apko, fix stuff, use tarfs
This pulls in the latest apko which has a bunch of breaking API changes.

Notably, digests are going to change because we fixed a bug with
checksums in our installed db file.

Also, go-apk now caches much more aggressively, so we expect to see a
large reduction in CPU and latency but an increase in how much disk the
cache consumes. This increased disk usage is somewhat offset by the fact
that we no longer unpack entire APKs to disk for every build. Instead,
we store a single copy of the uncompressed APK in the cache and read its
contents directly when constructing the image layer. I suspect on
ephemeral disks (our builds) the net result is ~cancelled out, given
that we no longer use as much temporary space. For really large builds,
the cache will probably take up a lot more space than the tempdirs did
previously. We may have to come up with an eviction strategy.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>