apko_tags: handle provides better (#293)
Today sometimes configs request some package, say, `nodejs`, which gets
provided by a package named `nodejs-21`. When it comes time to get the
tags for this config, we can't pass `nodejs` today because the resulting
config has no package with that name. And callers don't necessarily know
it's called `nodejs-21`, or else they would have requested that.

This makes `apko_tags` aware of this scenario and slightly less dumb,
and if it can't find a package named `nodejs`, will look for a package
named `nodejs-*` and assume that's the desired package.

---------

Signed-off-by: Jason Hall <jason@chainguard.dev>
Signed-off-by: James Petersen <jpetersenames@gmail.com>
Co-authored-by: James Petersen <jpetersenames@gmail.com>