Support multi-arch config outputs
Prior to this change, apko_config had a single "config" property that
contained an apko config, where the package list was the intersection
of the arch-specific package lists.

Most of the time, the package list is identical across architectures,
but architecture-specific dependencies (like libatomic) can lead to
divergence. Since we only include the intersection, we would omit these
architecture specific packages. When apko_build re-solved the locked
package list, it would add the missing packages back, so this worked
out, but was suboptimal.

We want to solve this mostly because this information being incomplete
prevents us from relying on it, but this also solves an annoying hole in
our reproducibility story, because the missing packages would not be
pinned to any version and could float.

In the future, we may want to upgrade these locked versions (e.g.
"foo=1.2.3-r4") to locked hashes (e.g. "foo><sha1hashgoeshere") for the
architecture-specific configs, but apko does not currently support that
and there would be quite a lot of prep work we'd need to do, but at
least we have a place to do it now.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>