Handle experiment tag and pointers in reflect
See https://github.com/chainguard-dev/apko/pull/1453

Since GID is optional and 0 is a valid value, we needed a way to
distinguish between unset and 0, so GID became a pointer.

This broke the reflect code (labeled HACK) that punted on pointers.

I've updated that code to handle pointers but also to ignore any fields
marked with `apko:"experimental"` because I don't want to pollute the
schema with them.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>