gitub: prevent CI drift when upgrading apko (#663)
When upgrading apko, which changes output format, we may have a drift
in digests of what imagetest provider calculates and what apko
provider calculates for imagetest sandboxes.

This leads to 404 of imagetest digests not found in the registry. Thus
test build images twice - once with stock apko & imagetest, and then
with the new apko provider under test => this should enable to use
either old or new digests for sandbox images.